requirements.txt
Flask==1.1.2
alembic==1.4.1
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.36.1
Flask-SQLAlchemy==2.4.1
Flask-Assets==2.0
Flask-Migrate==2.5.2
# Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-Session==0.3.1
six~=1.16.0
psycopg2-binary>=2.6.1
paho-mqtt==1.5.0
hbmqtt==0.9.5
aiohttp==3.6.2
aiotg==1.0.0
aiodns==2.0.0
google-api-python-client==1.7.12
aiofiles==0.4.0
async_timeout
aiomisc~=10.0.0
paho-mqtt
asyncpg==0.23.0
amqtt~=0.10.0a4
aioredis~=1.3.1
Werkzeug~=2.0.1
websockets~=9.1